Convert from uint8 to int64 after data is sent to GPU

+++ src/deep_taxon/nn/models/lit.py
@@ -135,7 +135,7 @@ class AbstractLit(LightningModule):
     def training_step(self, batch, batch_idx):
         seqs, target = batch
         output = self.forward(seqs)
-        loss = self._loss(output, target)
+        loss = self._loss(output, target.long())
         if self.hparams.classify:
             self.log(self.train_acc, self.accuracy(output, target), prog_bar=True)
         self.log_dict({self.train_loss: loss, 'time': self.step_time(), 'wall_time': time()})
@@ -153,7 +153,7 @@ class AbstractLit(LightningModule):
     def validation_step(self, batch, batch_idx):
         seqs, target = batch
         output = self(seqs)
-        loss = self._loss(output, target)
+        loss = self._loss(output, target.long())
         if self.hparams.classify:
             self.log(self.val_acc, self.accuracy(output, target), prog_bar=True)
         self.log_dict({self.val_loss: loss, 'time': self.step_time(), 'wall_time': time()})
@@ -167,7 +167,7 @@ class AbstractLit(LightningModule):
     def test_step(self, batch, batch_idx):
         seqs, target = batch
         output = self(seqs)
-        loss = self._loss(output, target)
+        loss = self._loss(output, target.long())
         self.log_dict({self.test_loss: loss, 'time': self.step_time(), 'wall_time': time()})
         return loss
